Removed dependency on hepdata-converter-ws-docker
The docker image needs to run the hepdata-converter-ws, so the dependency should be from hepdata-converter-ws-docker to here, not vice versa. I've therefore removed the docker submodule.
The travis tests now run inside the hepdata-converter docker image (installing the dependencies first via pip install - setup.py has been modified to include the test dependencies).
The -I flag in the pip install commands is to ignore the installed version of six from distutils, as was previously done in the docker build.